requirements.txt
# Зависимости
aiogram>=3.15.0
aiohttp         # Асинхронные запросы вместо requests
aiosqlite       # Асинхронная работа с SQLite
apscheduler     # Планировщик для рассылки в 7:00
python-dotenv
loguru          # Современное логирование
deep-translator